STX-166: continue

+++ Modules/Brand/Http/Middleware/BrandDBMiddleware.php
@@ -11,6 +11,7 @@ use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\DB;
 use Modules\Brand\Models\Brand;
+use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 
 final class BrandDBMiddleware
 {
@@ -21,10 +22,14 @@ final class BrandDBMiddleware
      * @param Closure(Request): (Response|RedirectResponse) $next
      * @return Response|RedirectResponse
      */
-    public function handle(Request $request, Closure $next)
-    {
+    public function handle(Request $request, Closure $next): Response|RedirectResponse
+    {dd('s');
+        $brand = Brand::where('slug', $request->header('Brand'))->first();
+
+        abort_if(! $request->user()->can('view', $brand), ResponseAlias::HTTP_FORBIDDEN);
+
         DB::purge('pgsql');
-        $brand = Brand::query()->find($request->header(['brand_id']));
+
         Config::set('database.connections.pgsql.database', $brand->slug);
 
         return $next($request);
